+++ package.json
@@ -7,19 +7,23 @@
     "dev:watch": "cross-env NODE_ENV=development nodemon -e ts -x 'ts-node-transpile-only src/index.ts'",
     "dev:integrated": "npm run devenv:build && concurrently npm:dev npm:devenv:deploy",
     "dev:follower": "npm run devenv:build && concurrently npm:dev npm:devenv:follower",
-    "test": "cross-env NODE_ENV=development jest --config ./jest.config.js --coverage --runInBand",
-    "test:rosetta": "cross-env NODE_ENV=development jest --config ./jest.config.rosetta.js --coverage --runInBand",
-    "test:rosetta-cli:data": "cross-env NODE_ENV=development STACKS_CHAIN_ID=0x80000000 jest --config ./jest.config.rosetta-cli-data.js --coverage --runInBand",
-    "test:rosetta-cli:construction": "cross-env NODE_ENV=development STACKS_CHAIN_ID=0x80000000 jest --config ./jest.config.rosetta-cli-construction.js --coverage --runInBand",
-    "test:bns": "cross-env NODE_ENV=development jest --config ./jest.config.bns.js --coverage --runInBand",
-    "test:microblocks": "cross-env NODE_ENV=development jest --config ./jest.config.microblocks.js --coverage --runInBand",
-    "test:tokens": "cross-env NODE_ENV=development jest --config ./jest.config.tokens.js --coverage --runInBand",
-    "test:watch": "cross-env NODE_ENV=development jest --config ./jest.config.js --watch",
-    "test:integration": "npm run devenv:deploy -- -d && cross-env NODE_ENV=development jest --config ./jest.config.js --coverage --no-cache --runInBand; npm run devenv:stop",
-    "test:integration:rosetta": "npm run devenv:deploy -- -d && cross-env NODE_ENV=development jest --config ./jest.config.rosetta.js --coverage --no-cache --runInBand; npm run devenv:stop",
-    "test:integration:bns": "npm run devenv:deploy -- -d && cross-env NODE_ENV=development jest --config ./jest.config.bns.js --coverage --no-cache --runInBand; npm run devenv:stop",
-    "test:integration:microblocks": "npm run devenv:deploy:pg -- -d && cross-env NODE_ENV=development jest --config ./jest.config.microblocks.js --coverage --no-cache --runInBand; npm run devenv:stop:pg",
-    "test:integration:tokens": "npm run devenv:deploy -- -d && cross-env NODE_ENV=development jest --config ./jest.config.tokens.js --coverage --no-cache --runInBand; npm run devenv:stop",
+    "test": "cross-env NODE_ENV=development jest --config ./tests/jest.config.js --coverage --runInBand",
+    "test:rosetta": "cross-env NODE_ENV=development jest --config ./tests/jest.config.rosetta.js --coverage --runInBand",
+    "test:rosetta-cli:data": "cross-env NODE_ENV=development STACKS_CHAIN_ID=0x80000000 jest --config ./tests/jest.config.rosetta-cli-data.js --coverage --runInBand",
+    "test:rosetta-cli:construction": "cross-env NODE_ENV=development STACKS_CHAIN_ID=0x80000000 jest --config ./tests/jest.config.rosetta-cli-construction.js --coverage --runInBand",
+    "test:bns": "cross-env NODE_ENV=development jest --config ./tests/jest.config.bns.js --coverage --runInBand",
+    "test:rpc": "cross-env NODE_ENV=development jest --config ./tests/jest.config.rpc.js --coverage --runInBand",
+    "test:btc-faucet": "cross-env NODE_ENV=development jest --config ./tests/jest.config.btc-faucet.js --coverage --runInBand",
+    "test:tokens": "cross-env NODE_ENV=development jest --config ./tests/jest.config.tokens.js --coverage --runInBand",
+    "test:watch": "cross-env NODE_ENV=development jest --config ./tests/jest.config.js --watch",
+    "test:integration": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml up\" \"cross-env NODE_ENV=development jest --config ./tests/jest.config.js --no-cache --runInBand; npm run devenv:stop:pg\"",
+    "test:integration:rosetta": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml up\" \"cross-env NODE_ENV=development jest --config ./tests/jest.config.rosetta.js --no-cache --runInBand; npm run devenv:stop\"",
+    "test:integration:rosetta-cli:data": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml up\" \"cross-env NODE_ENV=development STACKS_CHAIN_ID=0x80000000 jest --config ./tests/jest.config.rosetta-cli-data.js --no-cache --runInBand; npm run devenv:stop\"",
+    "test:integration:rosetta-cli:construction": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml up\" \"cross-env NODE_ENV=development STACKS_CHAIN_ID=0x80000000 jest --config ./tests/jest.config.rosetta-cli-construction.js --no-cache --runInBand\"",
+    "test:integration:bns": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml up\" \"cross-env NODE_ENV=development jest --config ./tests/jest.config.bns.js --no-cache --runInBand; npm run devenv:stop\"",
+    "test:integration:tokens": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml up\" \"cross-env NODE_ENV=development jest --config ./tests/jest.config.tokens.js --no-cache --runInBand; npm run devenv:stop\"",
+    "test:integration:rpc": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml up\" \"cross-env NODE_ENV=development jest --config ./tests/jest.config.rpc.js --no-cache --runInBand; npm run devenv:stop\"",
+    "test:integration:btc-faucet": "concurrently \"docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.bitcoind.yml up\" \"cross-env NODE_ENV=development jest --config ./tests/jest.config.btc-faucet.js --no-cache --runInBand; npm run devenv:stop\"",
     "generate:git-info": "node ./git-info-generator.mjs",
     "build": "npm run generate:git-info && rimraf ./lib && tsc -p tsconfig.build.json",
     "build:tests": "tsc -p tsconfig.json",
@@ -33,10 +37,9 @@
     "migrate": "node-pg-migrate -m src/migrations",
     "devenv:build": "docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml -f docker/docker-compose.dev.bitcoind.yml -f docker/docker-compose.dev.rosetta-cli.yml build --no-cache",
     "devenv:deploy": "docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml -f docker/docker-compose.dev.bitcoind.yml up",
-    "devenv:deploy:pg": "docker-compose -f docker/docker-compose.dev.postgres.yml up",
     "devenv:follower": "docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain-follower.yml up",
     "devenv:stop": "docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml -f docker/docker-compose.dev.bitcoind.yml down -v -t 0",
-    "devenv:stop:pg": "docker-compose -f docker/docker-compose.dev.postgres.yml kill",
+    "devenv:stop:pg": "docker-compose -f docker/docker-compose.dev.postgres.yml down -v -t 0",
     "devenv:logs": "docker-compose -f docker/docker-compose.dev.postgres.yml -f docker/docker-compose.dev.stacks-blockchain.yml -f docker/docker-compose.dev.bitcoind.yml logs -t -f"
   },
   "repository": {
@@ -175,7 +178,7 @@
     "@types/uuid": "7.0.5",
     "@typescript-eslint/eslint-plugin": "5.15.0",
     "@typescript-eslint/parser": "5.15.0",
-    "concurrently": "6.5.1",
+    "concurrently": "7.3.0",
     "docker-compose": "0.23.14",
     "eslint": "8.11.0",
     "eslint-config-prettier": "8.5.0",
