+++ tests/jest.config.bns.js
@@ -1,11 +1,11 @@
 module.exports = {
     preset: 'ts-jest',
-    rootDir: 'src',
+    rootDir: `${require('path').dirname(__dirname)}/src`,
     testMatch: ['<rootDir>/tests-bns/*.ts'],
     testPathIgnorePatterns: ['<rootDir>/tests-bns/setup.ts', '<rootDir>/tests-bns/teardown.ts'],
     collectCoverageFrom: ['<rootDir>/**/*.ts'],
     coveragePathIgnorePatterns: ['<rootDir>/tests*'],
-    coverageDirectory: '../coverage',
+    coverageDirectory: '<rootDir>/../coverage',
     globalSetup: '<rootDir>/tests-bns/setup.ts',
     globalTeardown: '<rootDir>/tests-bns/teardown.ts',
     testTimeout: 60000,
