+++ tests/jest.config.js
@@ -1,11 +1,11 @@
 module.exports = {
   preset: 'ts-jest',
-  rootDir: 'src',
+  rootDir: `${require('path').dirname(__dirname)}/src`,
   testMatch: ['<rootDir>/tests/**/*.ts'],
   testPathIgnorePatterns: ['<rootDir>/tests/setup.ts', '<rootDir>/tests/teardown.ts'],
   collectCoverageFrom: ['<rootDir>/**/*.ts'],
   coveragePathIgnorePatterns: ['<rootDir>/tests*'],
-  coverageDirectory: '../coverage',
+  coverageDirectory: '<rootDir>/../coverage',
   globalSetup: '<rootDir>/tests/setup.ts',
   globalTeardown: '<rootDir>/tests/teardown.ts',
   testTimeout: 20000,
