+++ tests/jest.config.rosetta-cli-construction.js
@@ -1,11 +1,11 @@
 module.exports = {
   preset: 'ts-jest',
-  rootDir: 'src',
+  rootDir: `${require('path').dirname(__dirname)}/src`,
   testMatch: ['<rootDir>/tests-rosetta-cli-construction/**/*.ts'],
   testPathIgnorePatterns: ['<rootDir>/tests-rosetta-cli-construction/setup.ts', '<rootDir>/tests-rosetta-cli-construction/teardown.ts'],
   collectCoverageFrom: ['<rootDir>/**/*.ts'],
   coveragePathIgnorePatterns: ['<rootDir>/tests*'],
-  coverageDirectory: '../coverage',
+  coverageDirectory: '<rootDir>/../coverage',
   globalSetup: '<rootDir>/tests-rosetta-cli-construction/setup.ts',
   globalTeardown: '<rootDir>/tests-rosetta-cli-construction/teardown.ts',
   testTimeout: 180000,
